platform-kernel: Fix to avoid untrusted pointer dereference
Currently, the compat ioctl call distinguishes itself using a global
flag. If a user sends a compat ioctl call followed by a normal ioctl
call, it may result in using a user passed address as a kernel address
in the fastrpcdriver. To address this issue, consider localizing the
compat flag for the ioctl call.

Signed-off-by: Anvesh Jain P <quic_ajainp@quicinc.com>
Change-Id: I746bced0fccc2fc761027cd5959545c8cf434588