platform-kernel: virtio_fastrpc: improve tx buf management
Currently, the code acquires the tx buf and proceeds without
returning it to the fastrpc BE when an exception occurs.
This results in a situation where, if an exception happens,
the tx buf is not returned promptly, eventually causing all
16 buffers to be occupied. To resolve this issue, we can
modify the process by storing the tx buf in an hlist upon
encountering an exception, and then prioritizing the
retrieval of buffers from this hlist for subsequent requests.
This solution ensures that the tx buf is efficiently recycled
even in exceptional cases, thereby preventing buffer exhaustion.

Change-Id: Icbd6a63f01f107888069d44ff6fafc2a0493ed04
Signed-off-by: mingzh <quic_mingzh@quicinc.com>
(cherry picked from commit 17b19efc97c70305a62ce57cb11a4ce2a242c81f)